Restrict workflow to main branch only
This commit modifies the .github/workflows/publish-filter.yml by restricting the generate and create filters steps to only run if the GitHub reference is the main branch. This change was made to refine the CI/CD pipeline, ensuring that the steps run only for changes to the main branch, thus preventing unnecessary runs for branches that don't yet need these procedures.