Update xargs to use dynamic processor count in publish-filter workflow
This commit adjusts the find/xargs call in the GitHub Actions workflow file "publish-filter.yml" to dynamically use the number of available processors. Previously the process was hardcoded to use four processors which wasn't ideal for systems with a different number of processors. Now, the 'nproc' command is used to assign the number of available processors to a variable 'NUM_PROCESSORS' and it is used as an argument value to '-P' option in xargs command. This change optimizes the execution of the temp_script.sh script by effectively utilizing the available resources in different systems.