Add support for allowed exception rules in AdBlock format (#225)
This update introduces the ability to handle allowed exception rules in AdBlock formatted files. In the AdBlock norm, '@@' at the beginning of the rule means it is an exception. The code is updated to recognize these rules, process them separately, and store them in a new HashSet<string> variable. This will distinguish which Domain Name System (DNS) requests to allow, enhancing the utility's flexibility and control over network traffic rules. The adjustment is accompanied by necessary changes in the tests, improving our project's robustness against future modifications.